+++ lobster/tools/trlc/trlc_tool.py
@@ -28,7 +28,6 @@ from trlc.errors import Message_Handler, TRLC_Error
 from trlc.trlc import Source_Manager
 
 from lobster.common.errors import PathError
-from lobster.common.io import lobster_write
 from lobster.common.items import Requirement
 from lobster.common.multi_file_input_tool import create_worklist, MultiFileInputTool
 
@@ -131,15 +130,8 @@ class LOBSTER_Trlc(MultiFileInputTool):
             if item:
                 items.append(item)
 
-        with open(options.out, "w", encoding="UTF-8") as fd:
-            # lobster-trace: trlc_req.Output_File
-            lobster_write(
-                fd=fd,
-                kind=Requirement,
-                generator="lobster-trlc",
-                items=items,
-            )
-        print(f"lobster-trlc: successfully wrote {len(items)} items to {options.out}")
+        # lobster-trace: trlc_req.Output_File
+        self._write_output(Requirement, options.out, items)
 
 
 def main(args: Optional[Sequence[str]] = None) -> int: