+++ tests_system/lobster_codebeamer/lobster_codebeamer_test_runner.py
@@ -2,7 +2,6 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import List, Optional, Union
 import yaml
-from tests_system.lobster_codebeamer.mock_server import PORT
 from tests_system.testrunner import TestRunResult, TestRunner
 from lobster.tools.codebeamer.codebeamer import main
 
@@ -17,9 +16,10 @@ class ConfigFileData:
     page_size: Optional[int] = None
     num_request_retry: Optional[int] = None
     retry_error_codes: Optional[List[int]] = None
+    verify_ssl: Optional[bool] = None
 
-    def set_default_root_token_out(self):
-        self.root = f"https://localhost:{PORT}"
+    def set_default_root_token_out(self, port: int):
+        self.root = f"https://localhost:{port}"
         self.token = "abcdef1234567890"
         self.out = "codebeamer.lobster"
 
@@ -38,6 +38,7 @@ class ConfigFileData:
         append_if_not_none("page_size", self.page_size)
         append_if_not_none("num_request_retry", self.num_request_retry)
         append_if_not_none("retry_error_codes", self.retry_error_codes)
+        append_if_not_none("verify_ssl", self.verify_ssl)
 
         with open(filename, mode='w', encoding="UTF-8") as file:
             yaml.dump(data, file)