+++ tests_system/lobster_trlc/test_inputs_zero.py
@@ -26,7 +26,7 @@ class ZeroInputTest(LobsterTrlcSystemTestCaseBase):
         completed_process = self._test_runner.run_tool_test()
         asserter = Asserter(self, completed_process, self._test_runner)
         asserter.assertNoStdErrText()
-        asserter.assertStdOutText(f"lobster-trlc: successfully wrote 0 items to "
+        asserter.assertStdOutText(f"lobster-trlc: wrote 0 items to "
                                   f"{OUT_FILE}\n")
         asserter.assertExitCode(0)
         asserter.assertOutputFiles()
@@ -48,7 +48,7 @@ class ZeroInputTest(LobsterTrlcSystemTestCaseBase):
         completed_process = self._test_runner.run_tool_test()
         asserter = Asserter(self, completed_process, self._test_runner)
         asserter.assertNoStdErrText()
-        asserter.assertStdOutText(f"lobster-trlc: successfully wrote 0 items to "
+        asserter.assertStdOutText(f"lobster-trlc: wrote 0 items to "
                                   f"{OUT_FILE}\n")
         asserter.assertExitCode(0)